Restricting edit & delete operations (293)

+++ 9_digital_marketplace/digital_marketplace/marketplace/views.py
@@ -57,6 +57,9 @@ def create_product(request):
 
 def product_edit(request, id):
     product = Product.objects.get(id=id)
+    if product.seller != request.user:
+        return redirect('invalid')
+    
     product_form = ProductForm(request.POST or None, request.FILES or None, instance=product)
     if request.method == "POST":
         if product_form.is_valid():
@@ -66,6 +69,9 @@ def product_edit(request, id):
 
 def product_delete(request, id):
     product = Product.objects.get(id=id)
+    if product.seller != request.user:
+        return redirect('invalid')
+    
     if request.method == "POST":
         product.delete()
         return redirect('index')
@@ -85,3 +91,6 @@ def register(request):
 
     form = UserRegistrationForm()
     return render(request, 'marketplace/register.html', { 'user_form': form })
+
+def invalid(request):
+    return render(request, 'marketplace/invalid.html')
